MVC Structure completed
The Dankify website has been restructured to work with the MVC template. The deletion of images has not been completed as this is done via an AJAX call, which will be implemented once the REST API has been exposed properly.

The API completes the first GET request, that needs to be send in the API test at the bottom of slide 8 for the assignment.