API Test works and AJAX delete button implemented
The whole assignment has been migrated to MVC. The API test works, but throws an error when it tries to perform a GET request on the image, that was added via a POST request. The API test refers to the test.php document. The document might be throwing the error due to it fetching the first image from the JSON data in the GET request instead of the image matching the image_id returned from the POST request.